% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-textmarkup.R
\name{updateTextMarkupInput}
\alias{updateTextMarkupInput}
\title{Change the value of a text markup input on the client}
\usage{
updateTextMarkupInput(session, inputId, labels = NULL, values = NULL,
  placeholders = NULL)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}

\item{inputId}{The id of the input object.}

\item{labels}{The labels to set for the input object.}

\item{values}{the values to set for the input object.}

\item{placeholders}{The placeholders to set for the input object.}
}
\description{
Change the value of a text markup input on the client.
}
\seealso{
\link{textMarkupInput}

Other input elements: \code{\link{textMarkupInput}}
}
\concept{input elements}
