% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-textmarkup.R
\name{updateTextMarkupInput}
\alias{updateTextMarkupInput}
\title{Change the value of a text markup input on the client}
\usage{
updateTextMarkupInput(session, inputId, labels = NULL, values = NULL,
  placeholders = NULL)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}

\item{inputId}{The id of the input object.}

\item{labels}{The labels to set for the input object.}

\item{values}{the values to set for the input object.}

\item{placeholders}{The placeholders to set for the input object.}
}
\description{
Change the value of a text markup input on the client.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
  library(shiny)
  library(shinyhighlightr)

  ui <- fluidPage(
    textMarkupInput(
      inputId = "highlight", labels = c("Expression", "Test Cases"),
      placeholder = c("Define Expression", "Define Test Cases"),
      values = list(
        "[1-2]{1}[0-9]{3}",
        c(
          "19th century", "1855",
          "between 1905 and 1910"
        )
      )
    ),

    div(
      class = "form-group shiny-input-container",
      actionButton("reset_expression", "Reset Expression"),
      actionButton("reset_test_cases", "Reset Test Cases")
    )
  )

  server <- function(input, output, session) {
    observeEvent(input$reset_expression, {
      updateTextMarkupInput(
        session, inputId = "highlight",
        labels = c("Expression Changed", NA),
        values = list("[1-2]{1}[7-8]{1}[0-9]{2}", NA),
        placeholders = c("Define Expression Changed", NA)
      )
    })

    observeEvent(input$reset_test_cases, {
      updateTextMarkupInput(
        session, inputId = "highlight",
        labels = c(NA, "Test Cases Changed"),
        values = list(NA, c("1750", "18th century")),
        placeholders = c(NA, "Define Test Cases Changed")
      )
    })
  }

  shinyApp(ui, server)
}

}
\seealso{
\link{textMarkupInput}
}
